fix: decoding of data in curve controller (#4)
* test: fix decoding of data

* feat: allow swapping using Eth

* modify curve controller setup

+++ src/curve/IStableSwapPool.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.10;
 
 interface IStableSwapPool {
-    function coins(uint i) external view returns (address);
+    function coins(uint256 i) external view returns (address);
 }
